plugins/logs: Move benchmarks into separate file
Just refactoring, no code changes.
